refactor(path): change list to lineStationResponses
